refactor: 필터 타입 카드뷰 비즈니스 로직 분리
